pyproject.toml
[project]
name = "al-dente"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex Berezovsky",email = "oxonkoalexey@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "openai (>=1.96.1,<2.0.0)",
    "markdown (>=3.8.2,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-stubs-ext (>=5.2.2,<6.0.0)",
    "mypy (>=1.17.0,<2.0.0)",
    "python-telegram-bot (>=22.3,<23.0)"
]

[tool.poetry]
packages = [{include = "al_dente", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
